chore: move new types to `explorer/types.ts`
