+++ packages/ui/client/components/explorer/ExplorerItem.vue
@@ -2,10 +2,10 @@
 import type { Task } from 'vitest'
 import { nextTick } from 'vue'
 import { runFiles, client } from '~/composables/client'
-import { caseInsensitiveMatch } from '~/utils/task'
 import { openedTreeItems, coverageEnabled } from '~/composables/navigation'
 import { hasFailedSnapshot } from '@vitest/ws-client'
-import {taskTree, UITaskTreeNode} from '~/composables/explorer/tree'
+import type { UITaskTreeNode} from '~/composables/explorer/types'
+import { taskTree } from '~/composables/explorer/tree'
 // import { useSearchTasks } from '~/composables/search'
 // import {UISuite, UITest} from '~/composables/explorer/types'
 // import { collapseUIEntry, expandUIEntry } from '~/composables/explorer'
